+++ core/remote/src/main/java/com/emotionstorage/remote/di/RemoteDataSourceModule.kt
@@ -1,10 +1,10 @@
 package com.emotionstorage.remote.di
 
-import com.emotionstorage.data.dataSource.DailyReportRemoteDataSource
-import com.emotionstorage.data.dataSource.MyPageRemoteDataSource
-import com.emotionstorage.data.dataSource.NotificationSettingRemoteDataSource
-import com.emotionstorage.data.dataSource.TimeCapsuleRemoteDataSource
-import com.emotionstorage.data.dataSource.UserRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.DailyReportRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.MyPageRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.NotificationSettingRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.TimeCapsuleRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.UserRemoteDataSource
 import com.emotionstorage.remote.dataSourceImpl.DailyReportRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.MyPageRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.NotificationSettingRemoteDataSourceImpl
