+++ data/src/main/java/com/emotionstorage/data/dataSource/remote/TimeCapsuleRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.data.dataSource
+package com.emotionstorage.data.dataSource.remote
 
 import com.emotionstorage.data.model.TimeCapsuleEntity
 import java.time.LocalDate
