+++ data/src/main/java/com/emotionstorage/data/repoImpl/NotificationSettingsRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.emotionstorage.data.repoImpl
 
-import com.emotionstorage.data.dataSource.NotificationSettingRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.NotificationSettingRemoteDataSource
 import com.emotionstorage.domain.common.DataState
 import com.emotionstorage.domain.model.NotificationSettings
 import com.emotionstorage.domain.repo.NotificationSettingRepository
