+++ data/src/main/java/com/emotionstorage/data/repoImpl/TimeCapsuleRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.emotionstorage.data.repoImpl
 
-import com.emotionstorage.data.dataSource.FavoriteResultEntity
-import com.emotionstorage.data.dataSource.TimeCapsuleRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.FavoriteResultEntity
+import com.emotionstorage.data.dataSource.remote.TimeCapsuleRemoteDataSource
 import com.emotionstorage.data.modelMapper.TimeCapsuleMapper
 import com.emotionstorage.domain.common.DataState
 import com.emotionstorage.domain.model.TimeCapsule
